refactor(server): clean xmpp client boundaries
