+++ server/crates/waddle-xmpp-client-wasm/src/lib.rs
@@ -49,13 +49,10 @@ use waddle_xmpp_client::xep::{
         ThreadStatusFilter,
     },
     xep0292::{build_fetch_vcard4_iq, build_publish_vcard4_iq, parse_pep_vcard4, VCard4},
-    xep0402::{
-        build_fetch_bookmarks_iq, build_publish_bookmark_iq, parse_bookmarks_response, BookmarkItem,
-    },
+    xep0402::{build_fetch_bookmarks_iq, parse_bookmarks_response, BookmarkItem},
     xep0492::{
-        build_fetch_dm_bookmarks_iq, build_publish_dm_bookmark_iq, build_retract_dm_bookmark_iq,
-        dm_notify_is_default, merge_notify, merge_notify_into_extensions, read_dm_bookmark_notify,
-        read_fallback_mode, read_rich_payload_opt_in, NotifyMode, NS_WADDLE_DM_BOOKMARKS,
+        build_fetch_dm_bookmarks_iq, read_dm_bookmark_notify, read_fallback_mode,
+        read_rich_payload_opt_in,
     },
 };
 use waddle_xmpp_client::{
