+++ server/crates/waddle-xmpp-client/src/messaging.rs
@@ -22,6 +22,7 @@ pub use builders::{
 };
 pub use call::{
     build_finish, build_finish_migrated, build_finish_with_options, build_finish_with_reason,
+    build_muji_jingle_content, build_muji_session_initiate, build_muji_session_terminate,
     build_proceed, build_propose, build_reject, build_reject_with_options, build_retract,
     build_retract_with_options, build_session_accept, build_session_initiate,
     build_session_terminate, jingle_reason_from_wire_name, jingle_reason_wire_name,
